Fix edge case with version selection
Previously the code would, if no version match was found, select the last version in the list (ie, the most-recently-released version). But if a developer posts their releases in a stupid order--for example, releases an update for 4.1, 4.2, and 4.3, but posts them to Moodle in the order 4.3, 4.2, 4.1, then the 4.1 release will be the most recent and that one will be selected. Now the list is sorted on the actual version string of the plugin to get the version with the largest version string rather than the most recent release. Thanks theme_boost_union!!!